codes/idxsurvey/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
PyMySQL==1.1.1
Werkzeug==3.1.3
